Fix the fix in windows cipd recipe module.
BUG=
TBR=sergiyb@chromim.org,akuegel@chromium.org

Review URL: https://codereview.chromium.org/1420203005

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@297404 0039d316-1c4b-4281-b951-d872f2087c98

+++ scripts/slave/recipe_modules/cipd/resources/bootstrap.py
@@ -173,8 +173,9 @@ def write_file(path, data):
   with os.fdopen(fd, 'w') as f:
     f.write(data)
   if not sys.platform in ('linux2', 'darwin'):
-    # On windows we should remove destination file first.
-    os.remove(path)
+    # On windows we should remove destination file if it exists.
+    if os.path.exists(path):
+      os.remove(path)
     # At this point we may crash, and it's OK, as next time we'll just
     # re-install CIPD from scratch.
   os.rename(temp_file, path)
